fix: inline list comment indentation
Comments require special lowering treatment to ensure they're indented
correctly, as they're attributed to being a child of a list node in the
AST when they're not actually part of the list in the source code, but
rather the following block.

Previously this used a heuristic to locate these misattributed comments
and correct them, but it didn't correctly account for comments that
appear within the middle of a set of nested list (specifically, at the
end of a list nested within an outer list).

This commit unifies comment handling using a simpler method: if the
comment wasn't indented in source, don't indent it now.