fix: bad AST when using hard tabs for spacing
When using hard tabs to indent a conjunction list, the generated AST may
fail to emit a conj_list node, specifically this block:

            ---- MODULE bananas ----
            Op == /\ A = 1
  tab ->       	  /\ B = 2
            ====

Emits the following AST nodes:

    operator_definition
    name: identifier
    def_eq
    definition: bound_infix_op
        lhs: conj_list
        conj_item
            bullet_conj
            bound_infix_op
            lhs: identifier_ref
            symbol: eq
            rhs: nat_number
        symbol: land
        rhs: bound_infix_op
        lhs: identifier_ref
        symbol: eq
        rhs: nat_number

Note the definition is for bounded_infix_op, with a list of 1 item (the
lhs) with an op of "land" for the second item.

            ---- MODULE Bananas ------
            Op == /\ A = 1
                  /\ B = 2
            =====

Compared to the above, that uses only whitespace for alignment:

    operator_definition
    name: identifier
    def_eq
    definition: conj_list
        conj_item
        bullet_conj
        bound_infix_op
            lhs: identifier_ref
            symbol: eq
            rhs: nat_number
        conj_item
        bullet_conj
        bound_infix_op
            lhs: identifier_ref
            symbol: eq
            rhs: nat_number

Where the definition node is correctly labelled as a "conj_list" kind,
with two conj_item nodes.