feat: token indentation rewriter
Prevents excessive indentation from one line to the next.

This adds a long desired post-processing pass over the format token list
before rendering to adjust indentation of nested blocks when all tokens
within the block are at 2 or more levels deeper than the previous line's
indentation ("excessively indented").

This pass results in more consistent indentation, while accounting for
tokens that appear "late" in the block that cause it to be properly
indented (a token eventually appears at N+1, not just N+2 or more).