test: add TLA Examples corpus
This expends the test corpus to 214 input files, and moves the corpus
test to be an integration test.

These files are most of the files from the TLA Examples repo:

    https://github.com/tlaplus/Examples

I've given a fair few of these files a quick review for formatting
issues, and have only found one oddity with pluscal translated output
which isn't too bad.

38 files currently cause spec breakage when formatted:

    * AllocatorImplementation.tla
    * BPConProof.tla
    * Bakery.tla
    * Boulanger.tla
    * ClientCentric.tla
    * ConfigFileGrammar.tla
    * EWD687a.tla
    * EWD840_json.tla
    * EWD998.tla
    * EWD998Chan.tla
    * EWD998ChanID.tla
    * EWD998ChanID_export.tla
    * EWD998ChanID_shiviz.tla
    * EWD998ChanTrace.tla
    * EWD998Chan_opts.tla
    * EWD998PCal.tla
    * EWD998_anim.tla
    * EWD998_opts.tla
    * EWD998_optsSC.tla
    * EWD998_proof.tla
    * HDiskSynod.tla
    * HanoiSeq.tla
    * InnerFIFO.tla
    * InnerSequential.tla
    * InnerSerial.tla
    * MCLiveWriteThroughCache.tla
    * MCWriteThroughCache.tla
    * PaxosCommit.tla
    * RealTime_SS.tla
    * SPA_Attack.tla
    * SPA_Attack_New.tla
    * Slush.tla
    * SumSequence.tla
    * TLAPlusGrammar.tla
    * VoteProof.tla
    * YoYoAllGraphs.tla
    * YoYoNoPruning.tla
    * YoYoPruning.tla

Some of these require only very minor fixes, while some have more
complex diffs that need investigating. A small number of bugs appear to
be responsible for the failures (bugs < files thankfully!).